Quero ser um sonhador
Neste mundo digital, fiz os enchants, pocoes, arrow switching, shop completion e enemy triggering on room enter

+++ The Twins/Assets/Script/Enemy Scripts/BulletScript.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class BulletScript : MonoBehaviour
 {
-    float bulletDamage;
+    int bulletDamage;
 
-    public void EnemyDamage(float damage)
+    public void EnemyDamage(int damage)
     {
         bulletDamage = damage;
     }
